light-sensor-reader/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
smbus